Better Mapping in UI
this update introduced the new all-pins map, as well as the functional forced refresh checkbox.
NOTE: if the geocoder cannot find a specific house, it tries the street, if it cant find that, it just does the city, that way there is a marker.
ALSO NOTE: the full map will always regenerate